// lib/gen/models/Tags.ts
/* tslint:disable */
/* eslint-disable */
/**
 * forum
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists } from './runtime';
/**
 * 
 * @export
 * @interface Tags
 */
export interface Tags {
    /**
     * 
     * @type {string}
     * @memberof Tags
     */
    id: TagsIdEnum;
}

/**
* @export
* @enum {string}
*/
export enum TagsIdEnum {
    PsEx = 'PS_EX',
    PsJoy = 'PS_JOY',
    PsEasy = 'PS_EASY',
    TimeSociety = 'TIME_SOCIETY',
    TimeStudent = 'TIME_STUDENT',
    Years10 = 'YEARS_10',
    Years20 = 'YEARS_20',
    YearsOv30 = 'YEARS_OV_30',
    PlayEasy = 'PLAY_EASY',
    PlayVeteran = 'PLAY_VETERAN',
    PlayHero = 'PLAY_HERO',
    JobAck = 'JOB_ACK',
    JobDef = 'JOB_DEF',
    JobHeal = 'JOB_HEAL'
}

export function TagsFromJSON(json: any): Tags {
    return TagsFromJSONTyped(json, false);
}

export function TagsFromJSONTyped(json: any, ignoreDiscriminator: boolean): Tags {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
    };
}

export function TagsToJSON(value?: Tags | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
    };
}